requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
django-jquery==3.1.0
psycopg2==2.9.2
python-decouple==3.5
sqlparse==0.4.2
